Add base color texture support!
- Need to use flags to add min/max filter
- Need to generate mimaps somehow -- not sure if bgfx supports this
- Need to support the other two textures that gltf supports

+++ bae/include/utils/Vertex.h
@@ -44,7 +44,7 @@ struct TexCoordVertex {
 	{
 		ms_declaration
 			.begin()
-			.add(bgfx::Attrib::TexCoord0, 2, bgfx::AttribType::Int16, true)
+			.add(bgfx::Attrib::TexCoord0, 2, bgfx::AttribType::Float, true)
 			.end();
 	}
 };
